+++ backend/src/test/java/com/festago/school/application/integration/AdminSchoolV1QueryServiceIntegrationTest.java
@@ -7,7 +7,7 @@ import static org.assertj.core.api.SoftAssertions.assertSoftly;
 import com.festago.common.exception.ErrorCode;
 import com.festago.common.exception.NotFoundException;
 import com.festago.common.querydsl.SearchCondition;
-import com.festago.school.application.AdminSchoolV1QueryService;
+import com.festago.school.application.v1.AdminSchoolV1QueryService;
 import com.festago.school.domain.School;
 import com.festago.school.domain.SchoolRegion;
 import com.festago.school.presentation.v1.dto.SchoolV1Response;
